Merge 1 37 to master

+++ enrichment/src/main/kotlin/com/rarible/protocol/union/enrichment/model/CollectionStatistics.kt
@@ -1,8 +1,10 @@
 package com.rarible.protocol.union.enrichment.model
 
+import java.math.BigInteger
+
 data class CollectionStatistics(
-    val itemCount: Long,
-    val ownerCount: Long,
+    val itemCount: BigInteger,
+    val ownerCount: BigInteger,
     val volumes: List<StatisticsPeriod>,
     val totalVolume: StatisticsValue,
     val floorPrice: StatisticsValue?,
